[fdbmonitor] Refactor and add CTest support
